Fixes and nearing the end
Fixes a parsing issue with some command lines by adjust the matcher pattern.
Removes the weird integer string array map and replaces it with a 2-dimensional string array.
Creates the table with proper ordering now and makes it far more digestible to the reader as well as correctly matches the command columns to their actual counterparts instead of one of the matches for that scene.